Update superagent to 0.15.7
This fixes app tests on node 0.10, which failed semi-deterministically
by timeout since the max sockets for the http agent would be reached
after a certain number of tests had run.

This can be fixed by increasing the max sockets (default is 5),
or disabling the agent as superagent chose to do.